+++ aqua/__main__.py
@@ -2,7 +2,7 @@ import logging
 
 import typer
 
-from aqua.parser import TitleParser, TocParser
+from aqua.parsers.parser import TrialParser
 
 logging.basicConfig(level=logging.INFO)
 
@@ -11,10 +11,8 @@ typer_app = typer.Typer(help='Aqua-Parser manager.')
 
 @typer_app.command(help='Start parser.')
 def run(filename: str):
-    toc_parser = TocParser()
-    title_parser = TitleParser()
-    toc_parser.parse(filename=filename)
-    title_parser.parse(filename=filename)
+    trial = TrialParser()
+    trial.parse(filename)
 
 
 if __name__ == '__main__':
