+++ aqua/schemas.py
@@ -1,19 +1,26 @@
+from dataclasses import dataclass
 from datetime import datetime
 from typing import Optional
 
-from pydantic import BaseModel
 
-
-class TrialTitle(BaseModel):
+@dataclass
+class TrialTitle:
     measurement_object: str
     project: str
     report_date: datetime
     responsible_person: str
 
 
-class Trials(BaseModel):
+@dataclass
+class TrialTOC:
     smd: str
     status: Optional[str]
     value_description: Optional[str]
     single_value: Optional[str]
     trial_object: Optional[str]
+
+
+@dataclass
+class Trial:
+    title: TrialTitle
+    toc: list[TrialTOC]
